+++ src/PDBTools.jl
@@ -2,6 +2,7 @@ module PDBTools
 
 import Dates
 import Downloads
+import JSON
 import PrecompileTools
 using Compat: @compat
 using Format: format, generate_formatter
@@ -11,7 +12,7 @@ using StaticArrays: SVector, SMatrix
 using TestItems: @testitem, @testmodule
 using OrderedCollections: OrderedDict
 using InlineStrings: InlineString, String1, String15
-import MolSimToolkitShared: center_of_mass, wrap, dihedral, get_atoms
+import MolSimToolkitShared: center_of_mass, wrap, dihedral, get_atoms, save, load
 
 # Version
 const VERSION = pkgversion(@__MODULE__)
@@ -43,11 +44,11 @@ export select_with_vmd
 # Tools
 export center_of_mass
 export moveto!
-export sasa_particles, sasa
+export sasa_particles, sasa, SASA
 export StandardAtomicRadii, CreamerUnitedAtomRadii, CustomAtomicRadii
 export read_unitcell, lattice_to_matrix, matrix_to_lattice
 export hydrogen_bonds
-export mvalue, CreamerDenaturedModel
+export mvalue, CreamerDenaturedModel, save, load
 @compat public parse_mvalue_server_sasa, gmx_delta_sasa_per_restype, delta_sasa_per_restype, creamer_delta_sasa
 
 # Custom residue and element definitions